Keep whitespaces at the extremities of codeblocks
Now that codeblocks tokens are correctly parsed, there is no need anymore
for parsing its content and then extracting its plaintext version.
This allows to save some computation and to preserve the whitespaces at
the beginning and end of a codeblock.

Previously, messages containing only an empty codeblock where not
displayed. Now the text of the codeblock defaults to '\n' so the message
is always displayed.