Use recovery code with all verification methods
The previous commit made one attempt at fixing an issue with verifying
keys, but was misguided: the issue at hand was not in attempting the
wrong method of authorization, but rather what was *passed* to the
method. Namely, the account password as opposed to the recovery phrase.
Regardless of terminology, the latter should be used. Certain code has
been restored, while the password parameter remains deleted.